+++ packages/create-app/templates/default-app/packages/client/src/App.tsx
@@ -2,17 +2,33 @@
  * © 2021 Thoughtworks, Inc.
  */
 
-import React, { ReactElement, useState } from 'react'
+import React, { ReactElement, useCallback, useState } from 'react'
 import { Container } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
-import { Routes, Route } from 'react-router-dom'
+import { Route, Routes, useNavigate } from 'react-router-dom'
 import EmissionsMetricsPage from './pages/EmissionsMetricsPage'
 import RecommendationsPage from './pages/RecommendationsPage/'
 import ErrorPage from './layout/ErrorPage'
 import HeaderBar from './layout/HeaderBar'
 import MobileWarning from './layout/MobileWarning'
+import { AxiosError } from 'axios'
+import { formatAxiosError } from './layout/ErrorPage/ErrorPage'
+import { ClientConfig } from './Config'
+import loadConfig from './ConfigLoader'
 
-function App(): ReactElement {
+interface AppProps {
+  config?: ClientConfig
+}
+export function App({ config = loadConfig() }: AppProps): ReactElement {
+  const navigate = useNavigate()
+  const onApiError = useCallback(
+    (e: AxiosError) => {
+      console.error(e)
+      navigate('/error', { state: formatAxiosError(e) })
+    },
+    [navigate],
+  )
+
   const [mobileWarningEnabled, setMobileWarningEnabled] = useState(
     window.innerWidth < 768,
   )
@@ -24,6 +40,7 @@ function App(): ReactElement {
   const useStyles = makeStyles(() => ({
     appContainer: {
       padding: 0,
+      height: 'calc(100vh - 65px)',
     },
   }))
 
@@ -43,13 +60,21 @@ function App(): ReactElement {
       <HeaderBar />
       <Container maxWidth={false} className={classes.appContainer}>
         <Routes>
-          <Route path="/" element={<EmissionsMetricsPage />} />
-          <Route path="/recommendations" element={<RecommendationsPage />} />
+          <Route
+            path="/"
+            element={
+              <EmissionsMetricsPage config={config} onApiError={onApiError} />
+            }
+          />
+          <Route
+            path="/recommendations"
+            element={
+              <RecommendationsPage config={config} onApiError={onApiError} />
+            }
+          />
           <Route path="/error" element={<ErrorPage />} />
         </Routes>
       </Container>
     </>
   )
 }
-
-export default App